Support disjunctive faceting on offline indices
I factorized the implementation into a disjunctive faceting helper. NOTE: Because it’s in the `helpers` package, it has to be public.
